[sc-733}[wip] Support MultiCriterion
